requirements.txt
alembic==1.12.1
aniso8601==9.0.1
APScheduler==3.10.4
asttokens==2.4.1
async_generator==1.10
attrs==23.1.0
backcall==0.2.0
backports.shutil_get_terminal_size==1.0.0
beautifulsoup4==4.12.2
better-exceptions==0.3.3
bs4==0.0.2
cacheout==0.15.0
cachetools==5.3.2
certifi>=2024.07.04
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
colored==1.4.4
cryptography>=42.0.4
cssselect==1.2.0
DBUtils==3.0.3
dateparser==1.1.8
decorator==5.1.1
executing==2.0.1
func_timeout==4.3.5
greenlet==3.0.1
h11==0.14.0
humanize==4.8.0
icalendar==5.0.11
idna>=3.7
influxdb==5.3.1
itsdangerous==2.1.2
iso-639==0.4.5
jedi==0.19.1
Jinja2>=3.1.4
jsonpath==0.82.2
jsonschema==4.19.2
loguru==0.7.2
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.7
openai==1.2.2
outcome==1.3.0.post0
parse==1.19.1
parse-torrent-title==2.7
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow>=10.3.0
proces==0.1.7
prompt-toolkit==3.0.40
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome>=3.19.1
Pygments==2.16.1
PyJWT==2.8.0
pymongo>=4.6.3
PyMySQL>=1.1.1
pyperclip==1.8.2
pypushdeer==0.0.3
pyquery==2.0.0
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python_hosts==1.0.4
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
redis==3.5.3
redis-py-cluster==2.1.3
regex==2023.10.3
requests>=2.32.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
selenium==4.9.1
six==1.16.0
slack-sdk==3.23.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
srt==3.5.3
terminal-layout==3.0.0
tqdm>=4.66.3
traitlets==5.13.0
trio==0.23.0
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver>=3.5.3
urllib3>=2.2.2
w3lib==2.0.1
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager>=4.0.1
websockets==12.0
Werkzeug>=3.0.0
wsproto==1.2.0
zhconv==1.4.3
bencodepy>=0.9.5
feapder[render]
qbittorrent-api
anitopy
plexapi
transmission-rpc
slack_bolt
fastapi[all]
uvicorn==0.30.6
jwt==1.3.1
python_jose==3.3.0
passlib[Bcrypt]==1.7.4
python-multipart==0.0.9
sse_starlette==2.1.3